Refactor set borehole publication status
- Move code which is only used for testing purposes
to the appropriate project.
- Improve code comments.
